update project configuration, fix imports
