+++ package.json
@@ -1,24 +1,18 @@
 {
-    "name": "jaculus-org-jaculus-tools-monorepo",
+    "name": "jaculus-tools-monorepo",
     "private": true,
     "packageManager": "pnpm@9.6.0",
     "type": "module",
     "scripts": {
         "build": "pnpm -r run build",
         "clean": "pnpm -r run clean",
-        "lint": "pnpm -r run lint",
+        "test": "vitest run",
+        "lint": "pnpm -r run lint && eslint --ext .ts ./unit",
         "lint:fix": "eslint . --fix",
         "format": "prettier --write .",
         "format:check": "prettier --check .",
         "format:all": "pnpm lint:fix && pnpm format",
-        "typecheck": "tsc -b --noEmit",
-        "test": "vitest run",
-        "dev:tools": "pnpm --filter jaculus-tools run dev",
         "jac": "pnpm --filter jaculus-tools exec jac",
-        "release:dry": "pnpm -r publish --dry-run",
-        "release": "pnpm -r publish --access public",
-        "publish:manual": "pnpm -r publish --access public --no-git-checks",
-        "publish:check": "pnpm -r publish --dry-run --access public",
         "prepare": "husky"
     },
     "devDependencies": {
@@ -27,6 +21,7 @@
         "chai-bytes": "^0.1.2",
         "eslint-config-prettier": "^10.1.8",
         "husky": "^9.1.7",
+        "jiti": "^2.5.1",
         "lint-staged": "^16.1.5",
         "prettier": "^3.6.2",
         "queue-fifo": "^0.2.5",
