+++ packages/device/tsconfig.json
@@ -2,9 +2,7 @@
     "extends": "../../tsconfig.base.json",
     "compilerOptions": {
         "outDir": "dist",
-        "rootDir": "src",
-        "composite": true
+        "rootDir": "src"
     },
-    "references": [{ "path": "../link" }, { "path": "../util" }],
     "include": ["src"]
 }
