+++ packages/link/src/mux.ts
@@ -1,4 +1,4 @@
-import { Logger } from "@jaculus/util/index.js";
+import { Logger } from "@jaculus/util";
 import { Encoder, Packetizer, Serializer } from "./encoders/interface.js";
 import { Consumer, Packet } from "./linkTypes.js";
 import { Duplex, OutputStream } from "./stream.js";
