+++ packages/link/tsconfig.json
@@ -2,9 +2,7 @@
     "extends": "../../tsconfig.base.json",
     "compilerOptions": {
         "outDir": "dist",
-        "rootDir": "src",
-        "composite": true
+        "rootDir": "src"
     },
-    "include": ["src"],
-    "references": [{ "path": "../util" }]
+    "include": ["src"]
 }
