+++ packages/tools/package.json
@@ -4,24 +4,21 @@
     "description": "Companion CLI for Jaculus",
     "license": "GPL-3.0-only",
     "type": "module",
-    "private": false,
     "bin": {
-        "jac": "dist/src/cli.js"
+        "jac": "dist/cli.js"
     },
-    "main": "dist/src/cli.js",
-    "module": "dist/src/cli.js",
-    "types": "dist/src/cli.d.ts",
+    "main": "dist/cli.js",
+    "types": "dist/cli.d.ts",
     "scripts": {
         "prebuild": "node -p \"'export const version = ' + JSON.stringify(require('./package.json').version) + ';'\" > src/meta/version.ts",
         "build": "tsc -b",
-        "dev": "tsc -b --watch",
         "clean": "rimraf dist tsconfig.tsbuildinfo",
         "lint": "eslint --ext .ts src"
     },
     "dependencies": {
         "@jaculus/device": "workspace:*",
         "@jaculus/link": "workspace:*",
-        "@jaculus/package": "workspace:*",
+        "@jaculus/firmware": "workspace:*",
         "@jaculus/project": "workspace:*",
         "@jaculus/util": "workspace:*",
         "chalk": "^5.4.1",
