--- unit/tsconfig.json
@@ -1,10 +1,0 @@
-{
-    "extends": "../tsconfig.base.json",
-    "include": ["./**/*.test.ts", "./**/*.ts"],
-    "compilerOptions": {
-        "types": ["vitest/globals", "chai"],
-        "noEmit": false,
-        "outDir": "../dist/unit",
-        "rootDir": "."
-    }
-}
